+++ src/components/layout/MainLayout.tsx
@@ -1,13 +1,13 @@
 import { Outlet } from 'react-router-dom';
 
 import Footer from '../Footer';
-import Header from './Header';
+import Header from './Header/Header';
 
 const MainLayout = () => {
   return (
     <>
       <Header />
-      <main className='pt-24'>
+      <main className="pt-24">
         <Outlet />
       </main>
       <Footer />
